feat: Update PDF processing to add folding effects
The update allows to process PDF by adding a paper fold background and a stamp for beneficiaries. Also includes changes to the `replace` function in `classes.py` to add the `folded` and `beneficiario` parameters to allow for these new features. Also a new constant `DATA_FOLDED` was created to point to the required images destination. Minor changes were done too in the`ruff.xml` config file and `__init__.py`.

Reason: The new feature allows documents to have a more real look by adding the paper fold effect and stamps. The changes to the `replace` function and the new constant are needed to implement this new functionality.